requirements.txt
cityscapesscripts==2.2.0
cycler==0.10.0
gdown==4.2.0
humanfriendly==9.2
kiwisolver==1.2.0
kornia==0.5.8
matplotlib==3.4.2
numpy==1.19.2
pandas==1.1.3
Pillow==8.3.1
prettytable==2.1.0
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.4.1
seaborn==0.11.1
torch==1.7.1+cu110
torchvision==0.8.2+cu110
tqdm==4.48.2
typing-extensions==3.7.4.3
wcwidth==0.2.5
yapf==0.31.0
scipy==1.7.0
ftfy==6.0.1
opencv-python==4.5.1.48
setuptools==59.5.0
pillow==8.2.0
imageio==2.4.1
timm==0.8.3.dev0
regex
einops
git+https://github.com/facebookresearch/detectron2.git